Version_1/Backend/requirements.txt
fastapi
starlette_session
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
pydantic[email]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose
PyJWT
python-multipart
sendgrid
authlib
httpx
itsdangerous
langchain
vecs
supabase==2.4.6
langchain-community>=0.0.38
langchain-google-genai